+++ passes/S035/testdata/src/a/outside_package.go
@@ -1,0 +1,21 @@
+package a
+
+import (
+	"a/schema"
+)
+
+func foutside() {
+	_ = schema.Schema{
+		AtLeastOneOf: []string{
+			".invalidreference",
+		},
+	}
+
+	_ = map[string]*schema.Schema{
+		"name": {
+			AtLeastOneOf: []string{
+				".invalidreference",
+			},
+		},
+	}
+}