+++ passes/S037/testdata/src/a/alias.go
@@ -1,0 +1,21 @@
+package a
+
+import (
+	s "github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+)
+
+func falias() {
+	_ = s.Schema{
+		ExactlyOneOf: []string{
+			".invalidreference", // want "invalid ExactlyOneOf attribute reference"
+		},
+	}
+
+	_ = map[string]*s.Schema{
+		"name": {
+			ExactlyOneOf: []string{
+				".invalidreference", // want "invalid ExactlyOneOf attribute reference"
+			},
+		},
+	}
+}